requirements.txt
telethon>=1.41.1
aiosqlite>=0.20.0
redis>=5.0.1
motor>=3.3.2
python-dotenv>=1.0.1
asyncio-throttle>=1.0.2
cryptography>=42.0.5
python-dateutil>=2.9.0
ujson>=5.9.0